// prisma/generated/zod/inputTypeSchemas/PrimaryLiftsUncheckedCreateInputSchema.ts
import type { Prisma } from '@prisma/client';

import { z } from 'zod';

export const PrimaryLiftsUncheckedCreateInputSchema: z.ZodType<Prisma.PrimaryLiftsUncheckedCreateInput> = z.object({
  id: z.string().cuid().optional(),
  createdAt: z.coerce.date().optional(),
  name: z.string(),
  creadedBy: z.string(),
  isDeleted: z.boolean().optional(),
  flield1: z.string().optional().nullable(),
  flield2: z.string().optional().nullable(),
  flield3: z.string().optional().nullable()
}).strict();

export default PrimaryLiftsUncheckedCreateInputSchema;